Extract config.py to break circular import between agent and reasons
Shared path/logging infrastructure (get_repo_root, get_sdlc_dir, log,
etc.) now lives in config.py. Both agent.py and reasons.py import from
it with no circular dependency. agent.py can now do a top-level import
of get_reasons_db from reasons.py instead of a fragile lazy import.
